% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_boxploth}
\alias{draw_key_crossbarh}
\alias{draw_key_hpath}
\alias{draw_key_pointrangeh}
\title{Horizontal key drawing functions}
\usage{
draw_key_hpath(data, params, size)

draw_key_pointrangeh(data, params, size)

draw_key_crossbarh(data, params, size)

draw_key_boxploth(data, params, size)
}
\value{
A grid grob.
}
\description{
Horizontal key drawing functions
}

