% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-ggproto.R
\name{flip_ggproto}
\alias{flip_geom}
\alias{flip_ggproto}
\alias{flip_stat}
\title{Flip ggproto objects}
\usage{
flip_geom(.gg, ...)

flip_stat(.gg, ...)
}
\arguments{
\item{.gg}{A Stat, Geom or Position ggproto.}

\item{...}{Methods to be overridden.}
}
\description{
These functions flip standard ggproto components to horizontal
orientation.
}

