% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-barh.R
\name{geom_barh}
\alias{geom_barh}
\title{Bars, rectangles with bases on y-axis}
\usage{
geom_barh(mapping = NULL, data = NULL, stat = "count",
  position = "stack", ..., width = NULL, binwidth = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Horizontal version of \code{\link[ggplot2]{geom_bar}()}.
}

