% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-histogramh.R
\name{geom_histogramh}
\alias{geom_histogramh}
\title{Horizontal histograms and frequency polygons.}
\usage{
geom_histogramh(mapping = NULL, data = NULL, stat = "bin",
  position = "stack", ..., binwidth = NULL, bins = NULL, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Horizontal version of \code{\link[ggplot2]{geom_histogram}()}.
}

