% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-violinh.R
\name{geom_violinh}
\alias{geom_violinh}
\title{Horizontal violin plot.}
\usage{
geom_violinh(mapping = NULL, data = NULL, stat = "ydensity",
  position = "dodge", ..., draw_quantiles = NULL, trim = TRUE,
  scale = "area", na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\description{
Horizontal version of \code{\link[ggplot2]{geom_violin}()}.
}

