% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-ggproto.R
\name{ggflipped}
\alias{ggflipped}
\title{Clone a ggproto object to be flipped}
\usage{
ggflipped(gg, ...)
}
\arguments{
\item{gg}{A ggproto object.}

\item{...}{Methods to be overridden.}
}
\description{
Clone a ggproto object and append \code{h} to its class name.
}

