% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerh.R
\name{layerh}
\alias{layerh}
\title{Create a flipped layer}
\usage{
layerh(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{layer}()}.}
}
\description{
The layer, geom and stat will be flipped to horizontal orientation.
}

