% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summaryh.R
\name{mean_se_h}
\alias{mean_se_h}
\title{Calculate mean and standard error}
\usage{
mean_se_h(x, mult = 1)
}
\arguments{
\item{x}{numeric vector}

\item{mult}{number of multiples of standard error}
}
\value{
A data frame with columns \code{x}, \code{xmin}, and \code{xmax}.
}
\description{
For use with \code{\link{stat_summaryh}}. Corresponding function for
vertical geoms is \code{\link[ggplot2]{mean_se}()}
}
\examples{
x <- rnorm(100)
mean_se_h(x)
}
