% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw.R
\name{draw_key}
\alias{draw_key}
\alias{draw_key_hpath}
\alias{draw_key_pointrangeh}
\alias{draw_key_crossbarh}
\alias{draw_key_boxploth}
\title{Horizontal key drawing functions}
\usage{
draw_key_hpath(data, params, size)

draw_key_pointrangeh(data, params, size)

draw_key_crossbarh(data, params, size)

draw_key_boxploth(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Horizontal key drawing functions
}
