% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-ggproto.R
\name{flip_ggproto}
\alias{flip_ggproto}
\title{Flip a ggproto object}
\usage{
flip_ggproto(gg)
}
\arguments{
\item{gg}{A ggproto object.}
}
\description{
S3 generic for flipping a ggproto object to horizontal orientation.
}

