% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-boxploth.R
\name{geom_boxploth}
\alias{geom_boxploth}
\title{Horizontal box and whiskers plot.}
\usage{
geom_boxploth(mapping = NULL, data = NULL, stat = "boxplot",
  position = "dodge", ..., outlier.colour = NULL, outlier.color = NULL,
  outlier.shape = 19, outlier.size = 1.5, outlier.stroke = 0.5,
  notch = FALSE, notchwidth = 0.5, varwidth = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Horizontal version of \code{\link[ggplot2]{geom_boxplot}()}.
}

