% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-crossbarh.R, R/geom-errorbarh.R, R/geom-linerangeh.R, R/geom-pointrangeh.R
\name{geom_crossbarh}
\alias{geom_crossbarh}
\alias{geom_errorbarh}
\alias{geom_linerangeh}
\alias{geom_pointrangeh}
\title{Horizontal intervals: lines, crossbars & errorbars.}
\usage{
geom_crossbarh(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., fatten = 2.5, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_errorbarh(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_linerangeh(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

geom_pointrangeh(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., fatten = 4, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Horizontal versions of \code{\link[ggplot2]{geom_linerange}()},
\code{\link[ggplot2]{geom_pointrange}()},
\code{\link[ggplot2]{geom_errorbar}()} and
\code{\link[ggplot2]{geom_crossbar}()}.
}

