% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-crossbarh.R, R/geom-errorbarh.R,
%   R/geom-linerangeh.R, R/geom-pointrangeh.R
\name{geom_crossbarh}
\alias{geom_crossbarh}
\alias{geom_errorbarh}
\alias{geom_linerangeh}
\alias{geom_pointrangeh}
\title{Horizontal intervals: lines, crossbars & errorbars.}
\usage{
geom_crossbarh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 2.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_errorbarh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_linerangeh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_pointrangeh(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  fatten = 4,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{fatten}{A multiplicative factor used to increase the size of the
middle bar in \code{geom_crossbar()} and the middle point in
\code{geom_pointrange()}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Horizontal versions of \code{\link[ggplot2]{geom_linerange}()},
\code{\link[ggplot2]{geom_pointrange}()},
\code{\link[ggplot2]{geom_errorbar}()} and
\code{\link[ggplot2]{geom_crossbar}()}.
}
\section{Aesthetics}{

\code{geom_crossbarh()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{xmin}}
  \item \strong{\code{xmax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


\code{geom_errorbarh()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{y}}
  \item \strong{\code{xmin}}
  \item \strong{\code{xmax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
  \item \code{width}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


\code{geom_linerangeh()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{y}}
  \item \strong{\code{xmin}}
  \item \strong{\code{xmax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.


\code{geom_pointrangeh()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \strong{\code{xmin}}
  \item \strong{\code{xmax}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{fill}
  \item \code{group}
  \item \code{linetype}
  \item \code{shape}
  \item \code{size}
  \item \code{stroke}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\section{Different between ggplot2 and ggstance}{


`ggplot2::geom_errorbarh()` uses the `height` aesthetic. The
ggstance version uses the `width` aesthetic. This is for
consistency with the direction of the geom and other ggstance
functions. You can still supply `height` for compatibility.
}

