% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstance.R, R/positionv.R
\docType{data}
\name{ggstance-ggproto}
\alias{PositionDodgev}
\alias{PositionFillv}
\alias{PositionJitterdodgev}
\alias{PositionStackv}
\alias{ggstance-ggproto}
\title{Base ggproto classes for ggstance}
\description{
Base ggproto classes for ggstance
}
\seealso{
ggplot2::ggproto
}
\keyword{datasets}
\keyword{internal}

