% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-barh.R, R/geom-boxploth.R, R/geom-crossbarh.R, R/geom-errorbarh.R, R/geom-linerangeh.R, R/geom-pointrangeh.R, R/geom-violinh.R, R/ggstance.R, R/position-dodgev.R, R/position-jitterdodgev.R, R/position-stackv.R, R/stat-binh.R, R/stat-boxploth.R, R/stat-counth.R, R/stat-xdensity.R
\docType{data}
\name{GeomBarh}
\alias{GeomBarh}
\alias{GeomBoxploth}
\alias{GeomCrossbarh}
\alias{GeomErrorbarh}
\alias{GeomLinerangeh}
\alias{GeomPointrangeh}
\alias{GeomViolinh}
\alias{PositionDodgev}
\alias{PositionFillv}
\alias{PositionJitterdodgev}
\alias{PositionStackv}
\alias{StatBinh}
\alias{StatBoxploth}
\alias{StatCounth}
\alias{StatXdensity}
\alias{ggstance-ggproto}
\title{Base ggproto classes for ggstance}
\description{
Base ggproto classes for ggstance
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}

