% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodgev.R, R/position-jitterdodgev.R, R/position-stackv.R, R/position.R
\name{position_dodgev}
\alias{position-vertical}
\alias{position_dodgev}
\alias{position_fillv}
\alias{position_jitterdodgev}
\alias{position_stackv}
\title{Vertical Positions}
\usage{
position_dodgev(height = NULL)

position_jitterdodgev(jitter.height = NULL, jitter.width = 0,
  dodge.height = 0.75)

position_stackv()

position_fillv()
}
\arguments{
\item{height}{Dodging height, when different to the height of the individual
elements. This is useful when you want to align narrow geoms with taller
geoms.}

\item{jitter.height}{degree of jitter in y direction. Defaults to 0.}

\item{jitter.width}{degree of jitter in x direction. Defaults to 40\% of the
resolution of the data.}

\item{dodge.height}{the amount to dodge in the y direction. Defaults to 0.75,
the default \code{position_dodgev()} height.}
}
\description{
Vertical versions of \code{\link[ggplot2]{position_dodge}()},
\code{\link[ggplot2]{position_jitterdodge}()},
\code{\link[ggplot2]{position_fill}()},
\code{\link[ggplot2]{position_stack}()},
}

